refactor(app): introduce Session facade and crate-level transcript boundary
The TUI reached directly into the runtime: `App` imported the
`everruns_core` event/message model, subscribed to the raw event
broadcast, and drove `run_turn`/`messages`/`events`/`execute_command`
by hand inside a ~180-line `start_turn`. That coupled rendering to
runtime internals and made every upstream event-model change ripple
into the UI.

Introduce two boundaries:

- `crate::transcript` — the single place that interprets `EventData`
  and `Message`. Owns the view-model types (`ChatLine`, `Author`,
  `ActivityStatus`, `StreamPreview`, `StreamKind`, `TurnEvent`,
  `DeltaRouter`) and all runtime-event/tool-result translation, moved
  out from under `app/`. Presentation (colors/labels) stays in
  `app::render`, so the translation layer is free of `ratatui`.

- `crate::session::Session` — the only owner of `RuntimeHandles` and
  the only caller of the runtime APIs. Exposes `run_turn`/`run_shell`
  (returning a `TurnHandle` with the `TurnEvent` stream + cancel),
  `replayed_lines`, and `execute_command` (returning a host-facing
  `CommandOutcome` so the runtime command type stays behind the facade).

`app/` no longer imports `everruns_core::events` or
`everruns_core::message::{Message, MessageRole}`, and no longer touches
the runtime directly outside the facade. `--print` (main.rs) and the
ACP bridge now call `crate::transcript` directly — the same shared
translation the TUI uses.

Adds an end-to-end facade test driving `Session::run_turn` over llmsim.
